requirements.txt
selenium
openai
python-dotenv
pyperclip